Separate the initialization of logger and tracer.
