+++ cmd/github-proxy/github-proxy.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
 	"github.com/sourcegraph/sourcegraph/internal/env"
+	"github.com/sourcegraph/sourcegraph/internal/logging"
 	"github.com/sourcegraph/sourcegraph/internal/tracer"
 )
 
@@ -60,6 +61,7 @@ var hopHeaders = map[string]struct{}{
 func main() {
 	env.Lock()
 	env.HandleHelpFlag()
+	logging.Init()
 	tracer.Init()
 
 	go func() {
