+++ cmd/query-runner/main.go
@@ -23,6 +23,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/eventlogger"
+	"github.com/sourcegraph/sourcegraph/internal/logging"
 	"github.com/sourcegraph/sourcegraph/internal/tracer"
 )
 
@@ -33,7 +34,7 @@ const port = "3183"
 func main() {
 	env.Lock()
 	env.HandleHelpFlag()
-
+	logging.Init()
 	tracer.Init()
 
 	go func() {
